pyproject.toml
[tool.poetry]
name = "socialsim4"
version = "0.1.0"
description = "SocialSim4 core simulation engine and web backend"
authors = ["SocialSim4 Team <dev@socialsim4.local>"]
packages = [{ include = "socialsim4", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
litestar = "^2.8.3"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.23"
psycopg = { extras = ["binary"], version = "^3.1.18" }
alembic = "^1.13.1"
bcrypt = "^5.0.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pydantic = "^2.4.2"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"
email-validator = "^2.1.0"
aiosqlite = "^0.20.0"
openai = "^1.58.1"
google-generativeai = "^0.7.2"
duckduckgo-search = "^7.3.2"
trafilatura = "^1.12.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
ruff = "^0.1.6"
mypy = "^1.6.0"

[tool.poetry.scripts]
socialsim4 = "socialsim4.cli:main"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
